Scale down mesh data. Fixed validation error by not destroying descriptor set layout early

+++ shader/simple.vert.glsl
@@ -6,9 +6,9 @@
 
 struct Vertex
 {
-    float vx, vy, vz;
+    float16_t vx, vy, vz;
     uint8_t nx, ny, nz, nw;
-    float tu, tv;
+    float16_t tu, tv;
 };
 
 layout(binding = 0) buffer readonly Vertices
